+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/EntityPanelBuilderTest.java
@@ -71,10 +71,10 @@ public class EntityPanelBuilderTest {
 		departmentPanelBuilder.detailPanel(employeePanelBuilder);
 
 		EntityPanel departmentPanel = departmentPanelBuilder.build(departmentModel);
-		assertEquals(departmentCaption, departmentPanel.caption().get());
+		assertEquals(departmentCaption, departmentPanel.caption());
 		assertThrows(IllegalArgumentException.class, () -> departmentPanel.detailPanel(Department.TYPE));
 		EntityPanel employeePanel = departmentPanel.detailPanel(Employee.TYPE);
-		assertEquals("empCaption", employeePanel.caption().get());
+		assertEquals("empCaption", employeePanel.caption());
 		assertEquals(1, departmentPanel.detailPanels().size());
 
 		assertEquals(departmentModel, departmentPanel.model());