+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityPanel.icon() added, used when panel is displayed in windows or tabbed panes.
 - EntityPanel.Builder.icon() added.
+- EntityPanel, caption, description and icon moved to Config, related changes.
 
 ## 0.17.37
 ### is.codion.common.core